ador_project/requirements.txt
# Core Dependencies
python-docx==1.1.0          # DOCX file parsing
spacy==3.7.2                # NER model (optional)
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.0/en_core_web_sm-3.7.0-py3-none-any.whl  # spaCy English model

# PDF Processing (for future LLM+RAG implementation)
pdfplumber==0.10.3          # PDF text extraction
pypdf2==3.0.1               # Alternative PDF parser
camelot-py[cv]==0.11.0      # Table extraction from PDFs

# Vector Database & Embeddings
chromadb==0.4.18            # Vector database for RAG
sentence-transformers==2.2.2 # Local embeddings

# LLM Integration
anthropic==0.7.8            # Claude API
openai==1.3.7               # OpenAI GPT-4 & embeddings
cohere==4.37                # Cohere reranking

# Web Framework (for API)
fastapi==0.104.1            # Modern API framework
uvicorn[standard]==0.24.0   # ASGI server
pydantic==2.5.0             # Data validation
python-multipart==0.0.6     # File upload support

# Task Queue (for async processing)
celery==5.3.4               # Distributed task queue
redis==5.0.1                # Message broker
kombu==5.3.4                # Celery dependencies

# Database
psycopg2-binary==2.9.9      # PostgreSQL driver
sqlalchemy==2.0.23          # ORM
alembic==1.13.0             # Database migrations

# Utilities
python-dotenv==1.0.0        # Environment variables
pyyaml==6.0.1               # Configuration files
requests==2.31.0            # HTTP client
aiohttp==3.9.1              # Async HTTP client

# Monitoring & Logging
prometheus-client==0.19.0   # Metrics export
structlog==23.2.0           # Structured logging

# Testing
pytest==7.4.3               # Unit testing
pytest-asyncio==0.21.1      # Async test support
pytest-cov==4.1.0           # Coverage reporting
httpx==0.25.2               # Test client for FastAPI

# Development Tools
black==23.12.0              # Code formatter
flake8==6.1.0               # Linter
mypy==1.7.1                 # Type checker
isort==5.13.2               # Import sorter

# Security
python-jose[cryptography]==3.3.0  # JWT tokens
passlib[bcrypt]==1.7.4      # Password hashing
cryptography==41.0.7        # Encryption utilities